Ensure acquired bootstrap resources can actually be used
